backend/requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0

# AI & Vision
rembg>=2.0.62
pillow>=10.1.0
opencv-python-headless>=4.8.0
numpy>=1.26.0
scikit-learn>=1.3.0
pytesseract>=0.3.10

# LLM Integration - Multiple Provider Support
# Install the one(s) you need:
openai>=1.3.0                    # Required for: OpenAI, Grok (xAI), Groq, Ollama
google-generativeai>=0.3.0       # Optional: For Google Gemini (free tier!)
httpx>=0.25.0
tenacity>=8.2.0

# Storage
boto3>=1.33.0
python-magic-bin>=0.4.14

# Utilities
python-dotenv>=1.0.0
structlog>=23.2.0
aiofiles>=23.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Optional: Background Tasks
# redis==5.0.1
# rq==1.15.1
# celery==5.3.4
